Update tests so that they run
* stubbed old counter code, left for reference
